Allow mods to control number of objective spawns (#463)
This change adds an event trigger to `XComTacticalMissionManager` that allows
listeners to override the number of objective spawns for a mission. This is
important if a mod needs to sync the number of objective spawns with some
other data, such as the rewards for the mission.

The event takes the form of:

```
{
  ID: OverrideObjectiveSpawnCount,
  Data: [ in BattleData BattleData, out int SpawnCount]
}
```