Add new events to customise unit status strings (#322)
Further to an earlier pull request, this change adds two extra events that
allow mods to customise unit status strings.

The first is 'OverridePersonnelStatus', which can override the status strings
determined by `GetPersonnelStatusSeparate()`. The event takes the form:

   {
      ID: OverridePersonnelStatus,
      Data: [out string Status, out string TimeLabel, out int TimeNum,
             out int State, out bool HideTime],
      Source: Unit
   }

The second event is 'OverridePersonnelStatusTime', which can override the
time parts of the status, for example by converting days to hours or vice
versa. The event takes the form:

   {
      ID: OverridePersonnelStatusTime,
      Data: [out string TimeLabel, out int TimeNum],
      Source: Unit
   }

This commit also introduces a new config variable -
CHHelpers.UseNewPersonnelStatusBehavior - that, when set, makes
GetPersonnelStatus() go through GetPersonnelStatusSeparate(). This is
to help ensure consistency between the two methods.

Finally, GetStatusStringsSeparate() and GetMentalStateStringsSeparate()
in XComGameState_Unit now trigger the 'OverridePersonnelStatusTime' event
so that listeners can override how times are displayed in those cases.