Add event for overriding allowed alert causes (#510)
Added an 'OverrideAllowedAlertCause' event to
`XCGS_AIUnitData.IsCauseAllowedForNonvisibleUnits()` that allows mods to
change which alert causes can put aliens on yellow alert outside of XCOM's
vision.

The event takes the form of:

  {
     ID: OverrideAllowedAlertCause,
     Data: [in int AlertCause, inout bool AllowThisCause]
  }

Note that it has no event source and `AlertCause` is an integer that matches
one of the `EAlertCause` enum values.