Add hook for SeesAlliesAlerted alert cause (#510)
The default behaviour for triggering the "SeesAlliesAlerted" alert doesn't
work well for yellow alert because of it's uncontrollable cascading effect.
This commit adds an 'OverrideSeesAlertedAllies' event that allows mods to
change that behaviour.

The event itself takes the form:

   {
      ID: OverrideSeesAlertedAllies,
      Data: [in XCGS_Unit UnitA, in XCGS_Unit UnitB, inout int AlertCause],
      Source: UnitA
   }

To disable the alert, listeners can simply set `AlertCause` to `eAC_None`.