Prevent psi zombies from dropping loot (#682)
Loot is supposed to be destroyed if a unit is killed with explosives,
but if such units are raised as psi zombies, those zombies will drop the
loot when *they* die.

This change clears the pending loot from units when they die, ensuring
that they can't drop the loot if they're raised from the dead.

Fixes #682.