Fix covert actions not being cleaned up (#819)
Issue #435 removed the unconditional clearing of a resistance faction's
covert actions when new ones were generated. This led to the
manifestation of an existing bug in a loop in
`CleanupFactionCovertActions()`, which meant that not all of the
existing covert actions were being removed.

This change simply replaces the loop with a reverse-traversal one that
can handle elements of the `CovertActions` array being removed during
iteration over that array.

Fixes #819.