Add support for grenade targeting
There was some extra camera rotation code in
`UITacticalHUD_AbilityContainer` that I missed first time round. That
code now also triggers the override event and honours the config
variable.

Also added a compatibility note to the event documentation so that mod
authors can be aware of the implications of overriding these classes.